feat(frontend): rename color token, add layer
