+++ frontend/src/pages/login/device-register/_components/passcode/passcode.stories.css.ts
@@ -3,7 +3,7 @@ import { style } from '@vanilla-extract/css';
 import { vars } from '@/features/theme';
 
 export const background = style({
-  background: vars.color.glassPrimary.background,
+  background: vars.color.glassSecondary.background,
   borderRadius: '10px',
   height: '300px',
